Revert "Merge updated onboarding view"
