+++ Nextcloud Cookbook iOS Client/Data/UserSettings.swift
@@ -86,6 +86,4 @@ class UserSettings: ObservableObject {
             return ""
         }
     }
-    
-    
 }
