+++ Nextcloud Cookbook iOS Client/Localizable.xcstrings
@@ -853,7 +853,6 @@
       }
     },
     "Entering the server address will open a web browser. Please follow the login instructions provided there. If the browser does not open, click the link 'Open in browser'\nAfter a successfull login, return to this application and press 'Validate'." : {
-      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -965,9 +964,6 @@
           }
         }
       }
-    },
-    "If the login button does not open your browser, copy the following link and paste it in your browser manually:" : {
-
     },
     "If you are interested in contributing to this project or simply wish to review its source code, we encourage you to visit the GitHub repository for this application." : {
       "localizations" : {
@@ -1239,9 +1235,6 @@
         }
       }
     },
-    "Login" : {
-
-    },
     "Login failed." : {
       "localizations" : {
         "de" : {
@@ -1285,9 +1278,6 @@
           }
         }
       }
-    },
-    "Make sure to enter the server address in the form 'example.com'. Currently, only servers using the 'https' protocol are supported." : {
-
     },
     "Missing recipe name." : {
       "localizations" : {
@@ -1510,7 +1500,6 @@
       }
     },
     "Open in browser" : {
-      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
@@ -1708,9 +1697,6 @@
         }
       }
     },
-    "Pull to refresh." : {
-
-    },
     "Same as Device" : {
       "localizations" : {
         "de" : {
@@ -1865,9 +1851,6 @@
         }
       }
     },
-    "Show help" : {
-
-    },
     "Submit" : {
       "localizations" : {
         "de" : {
@@ -1936,9 +1919,6 @@
           }
         }
       }
-    },
-    "The 'Login' button will open a web browser. Please follow the login instructions provided there.\nAfter a successful login, return to this application and press 'Validate'." : {
-
     },
     "The selected cookbook will open on app launch by default." : {
       "localizations" : {
